requirements.txt
streamlit
tensorflow==2.10.0
opencv-python
pillow
matplotlib
seaborn
scikit-learn

